refactor: 更改坐标计算方式
