+++ src/image/mod.rs
@@ -1,2 +1,1 @@
 pub mod maimai_best_50;
-pub mod utils;
