+++ src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod file;
 pub mod printer;
 pub mod simple_log;
+pub mod image;
